Fixed Mongo installer for RHEL

+++ Scripts/Deploy/Sherpa/Mongo/db_installer.sh
@@ -17,28 +17,50 @@
 
 
 
-sudo apt-key adv --keyserver hkp://keyserver.ubuntu.com:80 --recv EA312927
+if [ ! -f  "/etc/redhat-release" ];
+then
 
-# For Ubuntu 12.04
-echo "deb http://repo.mongodb.org/apt/ubuntu precise/mongodb-org/3.2 multiverse" | sudo tee /etc/apt/sources.list.d/mongodb-org-3.2.list
+    sudo apt-key adv --keyserver hkp://keyserver.ubuntu.com:80 --recv EA312927
 
-# For Ubuntu 14.04
-#echo "deb http://repo.mongodb.org/apt/ubuntu trusty/mongodb-org/3.2 multiverse" | sudo tee /etc/apt/sources.list.d/mongodb-org-3.2.list
+    # For Ubuntu 12.04
+    echo "deb http://repo.mongodb.org/apt/ubuntu precise/mongodb-org/3.2 multiverse" | sudo tee /etc/apt/sources.list.d/mongodb-org-3.2.list
 
-sudo apt-get update
-sudo apt-get install -y mongodb-org
-sudo service mongod start
+    # For Ubuntu 14.04
+    #echo "deb http://repo.mongodb.org/apt/ubuntu trusty/mongodb-org/3.2 multiverse" | sudo tee /etc/apt/sources.list.d/mongodb-org-3.2.list
 
-curl -sL https://deb.nodesource.com/setup_5.x | sudo -E bash -
-sudo apt-get install -y nodejs
-sudo apt-get install -y build-essential
+    sudo apt-get update
+    sudo apt-get install -y mongodb-org
+    sudo service mongod start
 
-npm install -g mongo-express
-cp /usr/lib/node_modules/mongo-express/config.default.js /usr/lib/node_modules/mongo-express/config.js
-npm install forever -g
-forever start /usr/lib/node_modules/mongo-express/app.js
+    curl -sL https://deb.nodesource.com/setup_5.x | sudo -E bash -
+    sudo apt-get install -y nodejs
+    sudo apt-get install -y build-essential
 
-export LC_ALL=C
-echo "export LC_ALL=C" >> /etc/environment
+    npm install -g mongo-express
+    cp /usr/lib/node_modules/mongo-express/config.default.js /usr/lib/node_modules/mongo-express/config.js
+    npm install forever -g
+    forever start /usr/lib/node_modules/mongo-express/app.js
+
+    export LC_ALL=C
+    echo "export LC_ALL=C" >> /etc/environment
+
+
+
+else
+
+echo "[mongodb-org-3.2]
+name=MongoDB Repository
+baseurl=https://repo.mongodb.org/yum/redhat/\$releasever/mongodb-org/3.2/x86_64/
+gpgcheck=1
+enabled=1
+gpgkey=https://www.mongodb.org/static/pgp/server-3.2.asc" > /etc/yum.repos.d/mongodb-org-3.2.repo
+
+sudo yum install -y mongodb-org
+mkdir -p /data/db
+#sudo service mongod start
+mongod --fork --logpath mongo.log
+
+fi
+
